Miscellaneous fixes.
* Refactores get_residues into two functions (based on the as_dict
parameter).
* Dataframe attribute is now calculated lazily. There's no need
to set the parameter explicitly in the constructor anymore; the
first time the `df` attribute is requested it is generated. If
the user wants to create it with different parameters, they
still can call generate_dataframe directly -- there is no loss
of functionality, we only add a bit of syntactic sugar.
* Aminoacid lists and maps were moved to constants and I added
a couple of helper functions.
* pdb_id is now inferred from the filename.
 + pending: should it be upper or lowercase?
* removed get_points_ function