Refactored alignment into its own module.
Uncoupled the alignment operation and from the conservation data
join. Hopefully in the future it can be used for other things
to add features in the dataframe.

read_conservation() is still in the Protein API to keep the
interface, but maybe later there will be a standalone function
to add features directly to the dataframe?