Added example notebook for graph models.
Something still doesn't feel right about having the graph both
in the generator and the protein object. I'll add an issue for it.

Also it might be a little bit annoying to have different node
identifiers for each graph model -- maybe we can find a standard
to agree upon. (xkcd/927) We should handle atoms, residues and
different kinds of representatives.