Fix file descriptor leak in vmd & persist IF
* Add option to persist getcontact's intermediate file, which would
avoid having to simulate everything again if rebooting a pipeline
* Bugfix: vmd-python leaks file descriptors!
compute_contacts calls gen_index_to_atom which calls
load_traj, and load_traj uses vmd.molecule.read..
you can check with lsof that there are file descriptors
that don't get cleaned up and they are related to stdout.
So, if graph_models.StaticContactGraphGenerator is run in
a script with >100 proteins to process, the script will fail
because of too many open FDs.
The solution I've implemented wraps the call in a new process,
which is easy since getcontacts outputs a file anyway. `lsof`
does not show an increasing amount of lines in time.